lib/hexdump: Use `size_t` for length parameter of `hexdump32()`
In the signature of the function `hexdump32()` it does not make sense to
represent a length, assumed to be positive, as a signed integer.
With this change, it is no longer necessary to cast a pointer to
unsigned long when passing it to `hexdump32()`.

The same change for the function `hexdump()` was done in commit
3dd0e72d [1].

	lib/hexdump: Take const void * and size_t as arguments

[1] http://review.coreboot.org/4575

Reviewed-on: http://review.coreboot.org/5646
Tested-by: build bot (Jenkins)

+++ src/include/lib.h
@@ -52,6 +52,6 @@ extern unsigned char _estack[];
 
 /* Defined in src/lib/hexdump.c */
 void hexdump(const void *memory, size_t length);
-void hexdump32(char LEVEL, const void *d, int len);
+void hexdump32(char LEVEL, const void *d, size_t len);
 
 #endif /* __LIB_H__ */
